refactor: split up test-utils util.rs (#12878)
