+++ crates/test-utils/src/lib.rs
@@ -21,9 +21,14 @@ pub mod fd_lock;
 mod filter;
 pub use filter::Filter;
 
+mod ext;
+pub use ext::ExtTester;
+
+mod prj;
+pub use prj::{TestCommand, TestProject};
+
 // Utilities for making it easier to handle tests.
 pub mod util;
-pub use util::{TestCommand, TestProject};
 
 mod script;
 pub use script::{ScriptOutcome, ScriptTester};
